# Use ndims in size to reduce allocations