--- a/Project.toml
+++ b/Project.toml
@@ -10,7 +10,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 Aqua = "0.5"
-CFITSIO = "1"
+CFITSIO = "1.2"
 Reexport = "0.2, 1.0"
 Tables = "1"
 julia = "1.3"