@jishnub 2021-07-29
After https://github.com/JuliaAstro/FITSIO.jl/pull/167/commits/73408ee2471443b3a4493b095f44d31a89e4cfd7 there are no allocations in `write`:

```julia
julia> f = FITS(tempname(), "w");

julia> a = Float64[1 3; 2 4];

julia> @btime write($f, $a);
  25.898 μs (0 allocations: 0 bytes)

julia> hdu = f[1];

julia> @btime write($hdu, $a);
  235.786 ns (0 allocations: 0 bytes)
```